+++ new_client/src/components/blogs/BlogPage.jsx
