+++ new_client/src/components/blogs/utils.js
@@ -10,3 +10,12 @@ export const markdownRender = (sanitizedInput) => {
     __html: marked(sanitizedInput),
   };
 };
+
+export const calculateReadingTime = (content) => {
+  const wordsPerMinute = 228;
+  const wordCount = content.replace(/[^\w ]/g, "").split(/\s+/).length;
+  const readingTimeInMinutes = Math.floor(wordCount / wordsPerMinute) + 1;
+  const readingTimeAsString = readingTimeInMinutes + " min";
+
+  return readingTimeAsString;
+};
